+++ src/routes/+layout.server.js
@@ -1,0 +1,6 @@
+export async function load({ request, setHeaders }) {
+	return {
+		// user: await db.getUser(request.headers.get("cookie"))
+		user: "Test"
+	};
+}
