+++ src/routes/auth/api/signup/+server.js
